e2e and integration tests for stock import
