+++ server/resources/templates/import-stock-template.csv
@@ -1,1 +1,2 @@
 inventory_group_name,inventory_code,inventory_text,inventory_type,inventory_unit,inventory_unit_price,inventory_cmm,stock_lot_label,stock_lot_quantity,stock_lot_expiration
+"Comprimé", "00001", "Quinine sulfate 500mg", "article", "comprimé", "0.02", "300", "QNN2020", "24000", "2020-12-31"
