+++ test/fixtures/bad-stock-to-import.csv
@@ -1,0 +1,11 @@
+"inventory_code","inventory_text","inventory_type","inventory_unit","inventory_unit_price","inventory_cmm","stock_lot_label","stock_lot_quantity","stock_lot_expiration"
+"","Abaisse langue en bois, 18*140mm, Boîte de 100 unités","article","Unité","1","629.17","CMPF0101","8700","2021-01-01"
+"","Abaisse langue en bois, 18*140mm, Boîte de 100 unités","article","Unité","1","629.17","022016C","6800","2021-02-01"
+"","Abaisse langue en bois, 18*140mm, Boîte de 100 unités","article","Unité","1","629.17","CMPF0601","16900","2021-05-01"
+"","Abaisse langue en bois, 18*140mm, Boîte de 100 unités","article","Unité","1","629.17","052016C","7900","2021-05-01"
+"","Abaisse langue en bois, 18*140mm, Boîte de 100 unités","article","Unité","1","629.17","CMPF43138","7500","2022-03-01"
+"","Abaisse langue en bois, 18*140mm, Boîte de 100 unités","article","Unité","1","629.17","NA","7600","2025-02-01"
+"","Acide Acetylsalicylique, 500mg, Tab, 1000, Vrac","article","Unité","1","49916.67","759170203","87000","2019-01-01"
+"","Acide Acetylsalicylique, 500mg, Tab, 1000, Vrac","article","Unité","1","49916.67","ASB17003","7000","2020-02-01"
+"","Acide Acetylsalicylique, 500mg, Tab, 1000, Vrac","article","Unité","1","49916.67","ASB17001","86000","2020-02-01"
+"","Acide Acetylsalicylique, 500mg, Tab, 1000, Vrac","article","Unité","1","49916.67","ASB17002","300","2020-03-01"
